+++ src/material/tooltip/testing/BUILD.bazel
@@ -20,14 +20,13 @@ filegroup(
 
 ng_test_library(
     name = "unit_tests_lib",
-    srcs = glob(
-        ["**/*.spec.ts"],
-        exclude = ["shared.spec.ts"],
-    ),
+    srcs = glob(["**/*.spec.ts"]),
     deps = [
-        ":harness_tests_lib",
         ":testing",
+        "//src/cdk/testing",
+        "//src/cdk/testing/testbed",
         "//src/material/tooltip",
+        "@npm//@angular/platform-browser",
     ],
 )
 
@@ -35,17 +34,5 @@ ng_web_test_suite(
     name = "unit_tests",
     deps = [
         ":unit_tests_lib",
-    ],
-)
-
-ng_test_library(
-    name = "harness_tests_lib",
-    srcs = ["shared.spec.ts"],
-    deps = [
-        ":testing",
-        "//src/cdk/testing",
-        "//src/cdk/testing/testbed",
-        "//src/material/tooltip",
-        "@npm//@angular/platform-browser",
     ],
 )
